Move stuff to lite
